refactor: rename app id to client id (#12)
Renaming the parameter appId to clientId. Using client id will make it
more obvious to users what the value corresponds to in the Unsplash API,
as this should really be a layer on top of that API.

BREAKING CHANGE: appId is no longer supported and users will need to use
clientId for the same value.